Package DiceThrowResult
