+++ main.go
@@ -1,19 +1,11 @@
 package main
 
 import (
-	chk "github.com/filippgerman/dnd/blob/e8e01b736093616d9bae87ab43058e0b492c8f0a/DiceChecking/DiceCheck"
+	chk "dnd/DiceChecking"
+	dth "dnd/GetThrowResult"
 	"github.com/gin-gonic/gin"
-	"math/rand"
 	"strconv"
-	"time"
 )
-
-func Get_DiceThrowResult(diceResult int) int {
-	rand.Seed(time.Now().UnixNano())
-	diceResult = (rand.Intn(diceResult))
-	diceResult += 1
-	return diceResult
-}
 
 func PostDice(c *gin.Context) {
 	diceSideNumber := c.Query("DiceRange")
@@ -23,7 +15,7 @@ func PostDice(c *gin.Context) {
 	}
 	chk.DiceCheckNil(&diceMaxNum)
 	c.JSON(200, gin.H{
-		"Result": Get_DiceThrowResult(diceMaxNum),
+		"Result": dth.Get_DiceThrowResult(diceMaxNum),
 	})
 }
 func main() {
